+++ models/player.rb
@@ -1,79 +1,83 @@
 class Player < Sequel::Model
-  def self.from_player_data(data)
-    player = Player.where(account_id: data.steam_id).first
-    if player.nil?
-      # Create new player if needed.
-      player = Player.create(
-        hive2_player_id: data.player_id,
-        account_id:      data.steam_id,
-        reinforced_tier: data.reinforced_tier
-      )
-    end
-
-    # Add new data point based on current data.
-    player_data = PlayerData.build_from_player_data(data, player_id: player.id)
-    player.update(current_player_data: player_data)
-
-    player
+  def self.get_or_create(account_id:)
+    Player.where(account_id: account_id).first ||
+      Player.create(account_id: account_id)
   end
 
   plugin :validation_helpers
   def validate
-    validates_presence [:account_id, :hive2_player_id]
+    validates_presence [:account_id]
   end
 
-  one_to_many :player_data,         class: :PlayerData
-  many_to_one :current_player_data, class: :PlayerData, key: :current_player_data_id
+  one_to_many :player_data_points
+  def _add_player_data_point(point)
+    # Important to do this first, as equality check also checks `player_id`.
+    point.player_id = id
+    point.relevant = current_player_data_point != point
+    point.save
+  end
+
+  many_to_one :current_player_data_point, class: :PlayerDataPoint, key: :current_player_data_point_id
   many_to_one :update_frequency
 
   def adagrad_sum
-    return nil unless current_player_data
-    current_player_data.adagrad_sum
+    return nil unless current_player_data_point
+    current_player_data_point.adagrad_sum
   end
 
   def alias
-    return nil unless current_player_data
-    current_player_data.alias
+    return nil unless current_player_data_point
+    current_player_data_point.alias
+  end
+
+  def hive_player_id
+    return nil unless current_player_data_point
+    current_player_data_point.hive_player_id
   end
 
   def experience
-    return nil unless current_player_data
-    current_player_data.experience
+    return nil unless current_player_data_point
+    current_player_data_point.experience
   end
 
   def level
-    return nil unless current_player_data
-    current_player_data.level
+    return nil unless current_player_data_point
+    current_player_data_point.level
+  end
+
+  def reinforced_tier
+    return nil unless current_player_data_point
+    current_player_data_point.reinforced_tier
   end
 
   def score
-    return nil unless current_player_data
-    current_player_data.score
+    return nil unless current_player_data_point
+    current_player_data_point.score
   end
 
   def skill
-    return nil unless current_player_data
-    current_player_data.skill
+    return nil unless current_player_data_point
+    current_player_data_point.skill
   end
 
   def time_total
-    return nil unless current_player_data
-    current_player_data.time_total
+    return nil unless current_player_data_point
+    current_player_data_point.time_total
   end
 
   def time_alien
-    return nil unless current_player_data
-    current_player_data.time_alien
+    return nil unless current_player_data_point
+    current_player_data_point.time_alien
   end
 
   def time_marine
-    return nil unless current_player_data
-    current_player_data.time_marine
+    return nil unless current_player_data_point
+    current_player_data_point.time_marine
   end
 
   def time_commander
-    return nil unless current_player_data
-    current_player_data.time_commander
+    return nil unless current_player_data_point
+    current_player_data_point.time_commander
   end
 
   def update_data(stalker: nil)
@@ -82,8 +86,14 @@ class Player < Sequel::Model
     begin
       Observatory::RateLimit.log_get_player_data(type: :background)
       data = stalker.get_player_data(account_id)
-      player_data = PlayerData.build_from_player_data(data, player_id: id)
-      update(current_player_data: player_data)
+      player_data = PlayerDataPoint.build_from_player_data_point(data)
+      add_player_data_point(player_data)
+      # Using `current_player_data_point` will *not* work, as that one checks
+      # whether the current and new object are equal, and only updates if they
+      # are not. 
+      # Due to us overwriting PlayerData#==, this would lead to non-relevant
+      # updates being discarded.
+      update(current_player_data_point_id: player_data.id)
 
       # Succesful updates will lead to reclassification.
       Resque.enqueue(Observatory::ClassifyPlayerUpdateFrequency, id)
@@ -118,23 +128,11 @@ class Player < Sequel::Model
   #
   # @param count [Fixnum] Number of entries which to return at most. `nil` in
   #   order to not limit entries.
-  # @return [Array<PlayerData>]
+  # @return [Array<PlayerDataPoint>]
   def recent_player_data(count = nil)
-    out = []
-    last_data = nil
-
-    player_data_dataset.order_by(Sequel.desc(:created_at)).each do |data|
-      if count && out.size >= count
-        return out
-      end
-
-      if last_data.nil? || last_data.time_total != data.time_total
-        out << data
-      end
-      last_data = data
-    end
-
-    out
+    player_data_points_dataset.where(relevant: true).
+      order_by(Sequel.desc(:created_at)).
+      limit(count)
   end
 
 #   def graph_time_played_total